==> 2. Analysis Code/5. Interpolation and Making results(All variables)/5. result_table_ym.ipynb <==
### **Code : Making result tables by year/month**
- Writer : Donghyeon Kim
- Date : 2022.09.13.
- 대상 : 통합 데이터(태양광 사용 & 미사용)

<hr>

### **Prior Settings**

# 라이브러리 설정
import os
import pandas as pd
import numpy as np
import openpyxl
import matplotlib.pyplot as plt
import seaborn as sns

# 사용자 함수 호출
# get_project_root : 파일의 상위-상위 경로 호출
# get_name_root : 모든 사용자 이름 호출
# get_name_root_use : 태양광 사용자 이름 호출
# get_name_root_not : 태양광 미사용자 이름 호출
from pack_utils import get_project_root, get_name_root, get_name_root_use, get_name_root_not, kw_dict, kw_value_dict

### **Data**
#### 1) Hour(1시간 단위)

# All Data
root = get_project_root()
folder_root = os.path.join(root, 'data_merge_wt_f')
file_name = os.path.join(folder_root, 'all_concat_hour.xlsx')
df_all_hour = pd.read_excel(file_name)

# Filtering
# (1) Using Solar
df_all_hour_use = df_all_hour[df_all_hour.type == 'use']

# (2) Not Using Solar
df_all_hour_not = df_all_hour[df_all_hour.type == 'not']

df_all_hour_use.info()

#### 2) Day(1일 단위)

# All Data
root = get_project_root()
folder_root = os.path.join(root, 'data_merge_wt_f')
file_name2 = os.path.join(folder_root, 'all_concat_day.xlsx')
df_all_day = pd.read_excel(file_name2)

# Filtering
# (1) Using Solar
df_all_day_use = df_all_day[df_all_day.type == 'use']

# (2) Not Using Solar
df_all_day_not = df_all_day[df_all_day.type == 'not']

df_all_day_use.info()

<hr>

### **1. 전력 소비량**
사용/미사용 가구 - 월별 평균 전력 소비량

# Result Dictionary
result = {}
result['설치가구(A)'] = []
result['미설치가구(B)'] = []
result['A/B'] = []

# Fill Dictionary
date_list = ['2021/4', '2021/5', '2021/6', '2021/7', '2021/8', '2021/9', '2021/10', '2021/11', '2021/12',
             '2022/1', '2022/2', '2022/3']
for i in range(len(date_list)):
    # 1) Year/Month 필터링
    df_all_day_use_f = df_all_day_use[df_all_day_use.ym == date_list[i]]
    df_all_day_not_f = df_all_day_not[df_all_day_not.ym == date_list[i]]
    
    # 2) 태양광 사용/미사용 가구별 값 합산
    use_consum = np.sum(df_all_day_use_f['전력소비량(kWh)'])
    not_consum = np.sum(df_all_day_not_f['전력소비량(kWh)'])
    
    # 3) 평균 구하기
    use_consum_mean = round(use_consum / len(df_all_day_use_f.owner.unique()), 1)
    not_consum_mean = round(not_consum / len(df_all_day_not_f.owner.unique()), 1)
    calc = round(use_consum_mean / not_consum_mean, 2)
    
    # 4) 값 넣기
    result['설치가구(A)'].append(use_consum_mean)
    result['미설치가구(B)'].append(not_consum_mean)
    result['A/B'].append(calc)
    
result = pd.DataFrame(result)
result.index = date_list
result

<hr>

### **2. 자가 발전량**
사용 가구(3kW) - 월별 발전량(전력 생산량)

# Result Dictionary
result = {}
result['발전량(kWh)'] = []
result['발전시간(h)'] = []
result['이용률(%)'] = []

# Fill Dictionary
date_list = ['2021/4', '2021/5', '2021/6', '2021/7', '2021/8', '2021/9', '2021/10', '2021/11', '2021/12',
             '2022/1', '2022/2', '2022/3']
for i in range(len(date_list)):
    # 1) Year/Month 필터링
    df_all_day_use_3kw = df_all_day_use[df_all_day_use['설비용량(kW)'] == '3kW']
    df_all_day_use_f = df_all_day_use_3kw[df_all_day_use_3kw.ym == date_list[i]]
    
    # 2) 태양광 사용/미사용 가구별 값 합산
    use_consum = np.sum(df_all_day_use_f['발전량(kWh)'])
    use_hour = np.sum(df_all_day_use_f['발전시간'])
    use_rate = np.sum(df_all_day_use_f['이용률'])

    # 3) 평균 구하기
    use_consum_mean = round(use_consum / len(df_all_day_use_f.owner.unique()), 1)
    use_hour_mean = round(use_hour / len(df_all_day_use_f), 1)
    use_rate_mean = round(use_rate / len(df_all_day_use_f), 1)

    # 4) 값 넣기
    result['발전량(kWh)'].append(use_consum_mean)
    result['발전시간(h)'].append(use_hour_mean)
    result['이용률(%)'].append(use_rate_mean)
    
result = pd.DataFrame(result)
result.index = date_list
result

<hr>

### **3. 잉여 전력량**
사용 가구 - 월별 평균 잉여 전력량

# Result Dictionary
result = {}
result['발전량(kWh)'] = []
result['잉여전력량(kWh)'] = []
result['잉여전력량/발전량'] = []

# Fill Dictionary
date_list = ['2021/4', '2021/5', '2021/6', '2021/7', '2021/8', '2021/9', '2021/10', '2021/11', '2021/12',
             '2022/1', '2022/2', '2022/3']
for i in range(len(date_list)):
    # 1) Year/Month 필터링
    df_all_day_use_f = df_all_day_use[df_all_day_use.ym == date_list[i]]
    
    # 2) 태양광 사용/미사용 가구별 값 합산
    use_consum = np.sum(df_all_day_use_f['발전량(kWh)'])
    use_export = np.sum(df_all_day_use_f['잉여전력량(kWh)'])
    
    # 3) 평균 구하기
    use_consum_mean = round(use_consum / len(df_all_day_use_f.owner.unique()), 1)
    use_export_mean = round(use_export / len(df_all_day_use_f.owner.unique()), 1)
    calc = round(use_export_mean / use_consum_mean, 2)
    
    # 4) 값 넣기
    result['발전량(kWh)'].append(use_consum_mean)
    result['잉여전력량(kWh)'].append(use_export_mean)
    result['잉여전력량/발전량'].append(calc)
    
result = pd.DataFrame(result)
result.index = date_list
result

<hr>

### **4. 수전 전력량**
사용/미사용 가구 - 월별 평균 수전 전력량

# Result Dictionary
result = {}
result['설치가구(A)'] = []
result['미설치가구(B)'] = []
result['A/B'] = []

# Fill Dictionary
date_list = ['2021/4', '2021/5', '2021/6', '2021/7', '2021/8', '2021/9', '2021/10', '2021/11', '2021/12',
             '2022/1', '2022/2', '2022/3']
for i in range(len(date_list)):
    # 1) Year/Month 필터링
    df_all_day_use_f = df_all_day_use[df_all_day_use.ym == date_list[i]]
    df_all_day_not_f = df_all_day_not[df_all_day_not.ym == date_list[i]]
    
    # 2) 태양광 사용/미사용 가구별 값 합산
    use_consum = np.sum(df_all_day_use_f['수전전력량(kWh)'])
    not_consum = np.sum(df_all_day_not_f['수전전력량(kWh)'])
    
    # 3) 평균 구하기
    use_consum_mean = round(use_consum / len(df_all_day_use_f.owner.unique()), 1)
    not_consum_mean = round(not_consum / len(df_all_day_not_f.owner.unique()), 1)
    calc = round(use_consum_mean / not_consum_mean, 2)
    
    # 4) 값 넣기
    result['설치가구(A)'].append(use_consum_mean)
    result['미설치가구(B)'].append(not_consum_mean)
    result['A/B'].append(calc)
    
result = pd.DataFrame(result)
result.index = date_list
result

<hr>

### **5. 자가소비율(SCR)**
사용 가구 - 월별 자가소비율

# Result Dictionary
result = {}
result['300W'] = []
result['3kW'] = []
result['18kW'] = []

# Fill Dictionary
date_list = ['2021/4', '2021/5', '2021/6', '2021/7', '2021/8', '2021/9', '2021/10', '2021/11', '2021/12',
             '2022/1', '2022/2', '2022/3']
for i in range(len(date_list)):
    # 1) Year/Month 필터링
    df_all_day_use_f = df_all_day_use[df_all_day_use.ym == date_list[i]]
    
    # 2) 설비용량 구분
    df_300w = df_all_day_use_f[df_all_day_use_f['설비용량(kW)'] == '300W']
    df_3kw = df_all_day_use_f[df_all_day_use_f['설비용량(kW)'] == '3kW']
    df_18kw = df_all_day_use_f[df_all_day_use_f['설비용량(kW)'] == '18kW']
    
    # 3) 발전용량별 자가소비율 합산
    scr_300w = np.sum(df_300w['자가소비율']) / len(df_300w)
    scr_3kw = np.sum(df_3kw['자가소비율']) / len(df_3kw)
    scr_18kw = np.sum(df_18kw['자가소비율']) / len(df_18kw)
    
    # 4) 평균 구하기
    scr_300w_mean = round(scr_300w * 100, 1)
    scr_3kw_mean = round(scr_3kw * 100, 1)
    scr_18kw_mean = round(scr_18kw * 100, 1)
    
    # 5) 값 넣기
    result['300W'].append(scr_300w_mean)
    result['3kW'].append(scr_3kw_mean)
    result['18kW'].append(scr_18kw_mean)
    
result = pd.DataFrame(result)
result.index = date_list
result

<hr>

### 6. **자가공급률(SSR)**
사용 가구 - 월별 자가공급률

# Result Dictionary
result = {}
result['300W'] = []
result['3kW'] = []
result['18kW'] = []

# Fill Dictionary
date_list = ['2021/4', '2021/5', '2021/6', '2021/7', '2021/8', '2021/9', '2021/10', '2021/11', '2021/12',
             '2022/1', '2022/2', '2022/3']
for i in range(len(date_list)):
    # 1) Year/Month 필터링
    df_all_day_use_f = df_all_day_use[df_all_day_use.ym == date_list[i]]
    
    # 2) 설비용량 구분
    df_300w = df_all_day_use_f[df_all_day_use_f['설비용량(kW)'] == '300W']
    df_3kw = df_all_day_use_f[df_all_day_use_f['설비용량(kW)'] == '3kW']
    df_18kw = df_all_day_use_f[df_all_day_use_f['설비용량(kW)'] == '18kW']
    
    # 3) 발전용량별 자가소비율 합산
    ssr_300w = np.sum(df_300w['자가공급률']) / len(df_300w)
    ssr_3kw = np.sum(df_3kw['자가공급률']) / len(df_3kw)
    ssr_18kw = np.sum(df_18kw['자가공급률']) / len(df_18kw)
    
    # 4) 평균 구하기
    ssr_300w_mean = round(ssr_300w * 100, 1)
    ssr_3kw_mean = round(ssr_3kw * 100, 1)
    ssr_18kw_mean = round(ssr_18kw * 100, 1)
    
    # 5) 값 넣기
    result['300W'].append(ssr_300w_mean)
    result['3kW'].append(ssr_3kw_mean)
    result['18kW'].append(ssr_18kw_mean)
    
result = pd.DataFrame(result)
result.index = date_list
result
